backend/requirements.txt
fastapi[standard]==0.115.12
passlib==1.7.4
psycopg2-binary==2.9.10
python-jose==3.5.0
SQLAlchemy==2.0.41